Move user model into separate file
